fix: 냉장고 페이지 재료 검색시 사용자 임의 입력 가능
