+++ src/pages/MyRefrigerator.tsx
@@ -20,8 +20,6 @@ export const MyRefrigerator = () => {
   );
   const [isSave, setIsSave] = useState(false);
 
-  // console.log(ingredientDetails);
-
   useEffect(() => {
     setIsSave(false);
     const newItems = addItemList.filter(
@@ -49,24 +47,13 @@ export const MyRefrigerator = () => {
       return;
     }
     setIsSave(!isSave);
-    // console.log(ingredientDetails);
   };
-
-  // const removeIngredient = (index: number) => {
-  //   setAddItemList(addItemList.filter((_, i) => i !== index));
-  //   setIngredientDetails(ingredientDetails.filter((_, i) => i !== index));
-  // };
 
   return (
     <>
       <BasicTitle title="나의 냉장고" />
       <Container>
-        <RefrigeSection></RefrigeSection>
-        <IngredientSearchForm
-          isRecipePageSearch={false}
-          addItemList={addItemList}
-          setAddItemList={setAddItemList}
-        />
+        <IngredientSearchForm addItemList={addItemList} setAddItemList={setAddItemList} />
         <Wrapper>
           {ingredientDetails.map((ingredient, index) => (
             <IngredientInfo
@@ -118,21 +105,4 @@ const Wrapper = styled.div`
   display: grid;
   grid-template-columns: 1fr 1fr;
   gap: 10px;
-`;
-
-const RefrigeSection = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  & > button {
-    width: 70px;
-  }
-
-  & > svg {
-    width: 25px;
-    height: 25px;
-    margin-right: 13px;
-    fill: ${(props) => props.theme.colors.darkGray};
-  }
 `;
